refactor: :recycle: rename content folder to data
