+++ package.json
@@ -39,7 +39,7 @@
 		"@babel/core": "^7.14.6",
 		"eslint": "7.2.0",
 		"eslint-config-airbnb": "18.2.1",
-		"eslint-config-next": "^11.0.1",
+		"eslint-config-next": "^11.1.0",
 		"eslint-config-node": "^4.1.0",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-plugin-import": "2.22.1",
