+++ pages/index.js
@@ -13,10 +13,10 @@ import NextLink from 'next/link'
 import Image from 'next/image'
 import Container from '../layout/container'
 import ProjectCard from '../components/projectCard'
-import projects from '../content/projects'
 import ContactForm from '../components/contactForm'
 import useColorModeSwitcher from '../utils/hooks/useColorModeSwitcher'
-import { twitter } from '../content/socials'
+import projects from '../data/projects'
+import { twitter } from '../data/socials'
 import ContentWrapper from '../layout/contentWrapper'
 
 export default function Home() {
